requirements.txt
streamlit
scikit-learn
numpy
pillow